requirements.txt
fastapi[all]
uvicorn==0.25.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
# pydantic==2.5.2

websockets==11.0.3
apscheduler==3.11.0
vaderSentiment==3.3.2
openai==1.59.9
mangum==0.17.0
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-openai==0.3.3
seaborn==0.12.2
scikit-image==0.22.0
scikit-learn==1.2.2
faiss-cpu==1.7.3

# JWT Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pyjwt==2.6.0

# PDF Processing
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
boto3==1.38.11
nltk==3.8.1